+++ src/otaclient/grpc/README.md
@@ -1,0 +1,3 @@
+# OTAClient gRPC Interface
+
+The implementation of the OTAClient gRPC interface.
