+++ src/otaclient/ota_core/README.md
@@ -1,0 +1,3 @@
+# OTAClient core
+
+The implementation of OTA request and OTA execution handling.
