+++ src/otaclient/ota_core/_updater.py
@@ -15,73 +15,26 @@
 
 from __future__ import annotations
 
-import atexit
-import errno
-import json
 import logging
-import multiprocessing.queues as mp_queue
 import os
 import shutil
-import threading
 import time
-from concurrent.futures import Future
-from functools import partial
-from hashlib import sha256
-from http import HTTPStatus
-from json.decoder import JSONDecodeError
 from pathlib import Path
-from queue import Empty, Queue
 from tempfile import TemporaryDirectory
-from typing import Any, Callable, NoReturn, Optional
-from urllib.parse import urlparse
-
-import requests.exceptions as requests_exc
-from requests import Response
 
 from ota_metadata.file_table.db import FileTableDBHelper
 from ota_metadata.file_table.utils import find_saved_fstable, save_fstable
-from ota_metadata.legacy2 import _errors as ota_metadata_error
-from ota_metadata.legacy2.metadata import (
-    OTAMetadata,
-    ResourceMeta,
-)
-from ota_metadata.utils.cert_store import (
-    CACertStoreInvalid,
-    CAChainStore,
-    load_ca_cert_chains,
-)
+from ota_metadata.legacy2.metadata import OTAMetadata, ResourceMeta
 from otaclient import errors as ota_errors
-from otaclient._download_resources import DownloadHelper
 from otaclient._status_monitor import (
-    OTAClientStatusCollector,
-    OTAStatusChangeReport,
     OTAUpdatePhaseChangeReport,
-    SetOTAClientMetaReport,
     SetUpdateMetaReport,
     StatusReport,
     UpdateProgressReport,
 )
-from otaclient._types import (
-    ClientUpdateControlFlags,
-    ClientUpdateRequestV2,
-    FailureType,
-    IPCRequest,
-    IPCResEnum,
-    IPCResponse,
-    MultipleECUStatusFlags,
-    OTAStatus,
-    UpdatePhase,
-    UpdateRequestV2,
-)
-from otaclient._utils import (
-    SharedOTAClientMetricsReader,
-    SharedOTAClientStatusWriter,
-    get_traceback,
-    wait_and_log,
-)
-from otaclient.boot_control import BootControllerProtocol, get_boot_controller
-from otaclient.client_package import OTAClientPackageDownloader
-from otaclient.configs._cfg_consts import CANONICAL_ROOT
+from otaclient._types import UpdatePhase
+from otaclient._utils import wait_and_log
+from otaclient.boot_control import BootControllerProtocol
 from otaclient.configs.cfg import cfg, ecu_info, proxy_info
 from otaclient.create_standby._common import ResourcesDigestWithSize
 from otaclient.create_standby.delta_gen import (
@@ -94,13 +47,13 @@ from otaclient.create_standby.delta_gen import (
 from otaclient.create_standby.resume_ota import ResourceScanner
 from otaclient.create_standby.update_slot import UpdateStandbySlot
 from otaclient.create_standby.utils import can_use_in_place_mode
-from otaclient.metrics import OTAMetricsData
 from otaclient_common import _env, human_readable_size, replace_root
-from otaclient_common.cmdhelper import ensure_mount, ensure_umount, mount_tmpfs
-from otaclient_common.common import ensure_otaproxy_start
-from otaclient_common.downloader import DownloaderPool
+from otaclient_common.cmdhelper import ensure_umount
 from otaclient_common.linux import fstrim_at_subprocess
 from otaclient_common.persist_file_handling import PersistFilesHandler
+
+from ._common import download_exception_handler
+from ._updater_base import OTAUpdateOperator
 
 logger = logging.getLogger(__name__)
 
@@ -109,10 +62,6 @@ WAIT_BEFORE_REBOOT = 6
 DOWNLOAD_STATS_REPORT_BATCH = 300
 DOWNLOAD_REPORT_INTERVAL = 1  # second
 
-OP_CHECK_INTERVAL = 1  # second
-HOLD_REQ_HANDLING_ON_ACK_REQUEST = 16  # seconds
-HOLD_REQ_HANDLING_ON_ACK_CLIENT_UPDATE_REQUEST = 4  # seconds
-WAIT_FOR_OTAPROXY_ONLINE = 3 * 60  # 3mins
 
 STANDBY_SLOT_USED_SIZE_THRESHOLD = 0.8
 
@@ -122,247 +71,7 @@ hold the OTA image metadata of standby slot itself.
 """
 
 
-class OTAClientError(Exception): ...
-
-
-def _download_exception_handler(_fut: Future[Any]) -> bool:
-    """Parse the exception raised by a downloading task.
-
-    This handler will raise OTA Error on exceptions that cannot(should not) be
-        handled by us. For handled exceptions, just let upper caller do the
-        retry for us.
-
-    Raises:
-        UpdateRequestCookieInvalid on HTTP error 401 or 403,
-        OTAImageInvalid on HTTP error 404,
-        StandbySlotInsufficientSpace on disk space not enough.
-
-    Returns:
-        True on succeeded downloading, False on handled exceptions.
-    """
-    if not (exc := _fut.exception()):
-        return True
-
-    try:
-        # exceptions that cannot be handled by us
-        if isinstance(exc, requests_exc.HTTPError):
-            _response = exc.response
-            # NOTE(20241129): if somehow HTTPError doesn't contain response,
-            #       don't do anything but let upper retry.
-            # NOTE: bool(Response) is False when status_code != 200.
-            if not isinstance(_response, Response):
-                return False
-
-            http_errcode = _response.status_code
-            if http_errcode in [HTTPStatus.FORBIDDEN, HTTPStatus.UNAUTHORIZED]:
-                raise ota_errors.UpdateRequestCookieInvalid(
-                    f"download failed with critical HTTP error: {exc.errno}, {exc!r}",
-                    module=__name__,
-                )
-            if http_errcode == HTTPStatus.NOT_FOUND:
-                raise ota_errors.OTAImageInvalid(
-                    f"download failed with 404 on some file(s): {exc!r}",
-                    module=__name__,
-                )
-
-        if isinstance(exc, OSError) and exc.errno == errno.ENOSPC:
-            raise ota_errors.StandbySlotInsufficientSpace(
-                f"download failed due to space insufficient: {exc!r}",
-                module=__name__,
-            )
-
-        # handled exceptions, let the upper caller do the retry
-        return False
-    finally:
-        del exc, _fut  # drop ref to exc instance
-
-
-class _OTAUpdateOperator:
-    """The base common class of OTA update logic."""
-
-    def __init__(
-        self,
-        *,
-        version: str,
-        raw_url_base: str,
-        cookies_json: str,
-        session_wd: Path,
-        ca_chains_store: CAChainStore,
-        upper_otaproxy: str | None = None,
-        ecu_status_flags: MultipleECUStatusFlags,
-        status_report_queue: Queue[StatusReport],
-        session_id: str,
-        metrics: OTAMetricsData,
-        shm_metrics_reader: SharedOTAClientMetricsReader,
-    ) -> None:
-        self.update_version = version
-        self.update_start_timestamp = int(time.time())
-        self.session_id = session_id
-        self._status_report_queue = status_report_queue
-        self._metrics = metrics
-        self._shm_metrics_reader = shm_metrics_reader
-
-        # ------ report INITIALIZING status ------ #
-        status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAUpdatePhaseChangeReport(
-                    new_update_phase=UpdatePhase.INITIALIZING,
-                    trigger_timestamp=self.update_start_timestamp,
-                ),
-                session_id=session_id,
-            )
-        )
-        self._metrics.initializing_start_timestamp = self.update_start_timestamp
-
-        status_report_queue.put_nowait(
-            StatusReport(
-                payload=SetUpdateMetaReport(
-                    update_firmware_version=version,
-                ),
-                session_id=session_id,
-            )
-        )
-        self._metrics.target_firmware_version = version
-
-        # ------ parse cookies ------ #
-        logger.debug("process cookies_json...")
-        try:
-            cookies = json.loads(cookies_json)
-            assert isinstance(
-                cookies, dict
-            ), f"invalid cookies, expecting json object: {cookies_json}"
-        except (JSONDecodeError, AssertionError) as e:
-            _err_msg = f"cookie is invalid: {cookies_json=}"
-            logger.error(_err_msg)
-            raise ota_errors.InvalidUpdateRequest(_err_msg, module=__name__) from e
-
-        # ------ parse upper proxy ------ #
-        logger.debug("configure proxy setting...")
-        self._upper_proxy = upper_otaproxy
-
-        # ------ mount session wd as a tmpfs ------ #
-        self._session_workdir = session_wd
-        session_wd.mkdir(exist_ok=True, parents=True)
-
-        # ------ init updater implementation ------ #
-        self.ecu_status_flags = ecu_status_flags
-
-        # ------ init variables needed for update ------ #
-        _url_base = urlparse(raw_url_base)
-        _path = f"{_url_base.path.rstrip('/')}/"
-        self.url_base = _url_base._replace(path=_path).geturl()
-
-        # ------ setup downloader ------ #
-        self._downloader_pool = _downloader_pool = DownloaderPool(
-            instance_num=cfg.DOWNLOAD_THREADS,
-            hash_func=sha256,
-            chunk_size=cfg.CHUNK_SIZE,
-            cookies=cookies,
-            # NOTE(20221013): check requests document for how to set proxy,
-            #                 we only support using http proxy here.
-            proxies={"http": upper_otaproxy} if upper_otaproxy else None,
-        )
-        self._download_helper = DownloadHelper(
-            downloader_pool=_downloader_pool,
-            max_concurrent=cfg.MAX_CONCURRENT_DOWNLOAD_TASKS,
-            download_inactive_timeout=cfg.DOWNLOAD_INACTIVE_TIMEOUT,
-        )
-
-        # ------ setup OTA metadata parser ------ #
-        self._ota_metadata = OTAMetadata(
-            base_url=self.url_base,
-            session_dir=self._session_workdir,
-            ca_chains_store=ca_chains_store,
-        )
-
-    def _handle_upper_proxy(self) -> None:
-        """Ensure the upper proxy is online before starting the local OTA update."""
-        if _upper_proxy := self._upper_proxy:
-            logger.info(
-                f"use {_upper_proxy} for local OTA update, "
-                f"wait for otaproxy@{_upper_proxy} online..."
-            )
-
-            # NOTE: will raise a built-in ConnnectionError at timeout
-            ensure_otaproxy_start(
-                _upper_proxy,
-                probing_timeout=WAIT_FOR_OTAPROXY_ONLINE,
-            )
-
-    def _process_metadata(self, only_metadata_verification: bool = False) -> None:
-        """Process the metadata.jwt file and report."""
-        _current_time = int(time.time())
-        self._status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAUpdatePhaseChangeReport(
-                    new_update_phase=UpdatePhase.PROCESSING_METADATA,
-                    trigger_timestamp=_current_time,
-                ),
-                session_id=self.session_id,
-            )
-        )
-        self._metrics.processing_metadata_start_timestamp = _current_time
-
-        logger.info("verify and download OTA image metadata ...")
-        try:
-            _condition = threading.Condition()
-            for _fut in self._download_helper.download_meta_files(
-                self._ota_metadata.download_metafiles(
-                    _condition,
-                    only_metadata_verification=only_metadata_verification,
-                ),
-                condition=_condition,
-            ):
-                _download_exception_handler(_fut)
-
-            _metadata_jwt = self._ota_metadata.metadata_jwt
-            assert _metadata_jwt, "invalid metadata jwt"
-
-            logger.info(
-                "ota_metadata parsed finished: \n"
-                f"total_regulars_num: {self._ota_metadata.total_regulars_num} \n"
-                f"total_regulars_size: {_metadata_jwt.total_regular_size}"
-            )
-
-            self._status_report_queue.put_nowait(
-                StatusReport(
-                    payload=SetUpdateMetaReport(
-                        image_file_entries=self._ota_metadata.total_regulars_num,
-                        image_size_uncompressed=_metadata_jwt.total_regular_size,
-                        metadata_downloaded_bytes=self._downloader_pool.total_downloaded_bytes,
-                    ),
-                    session_id=self.session_id,
-                )
-            )
-            self._metrics.ota_image_total_files_size = _metadata_jwt.total_regular_size
-            self._metrics.ota_image_total_regulars_num = (
-                self._ota_metadata.total_regulars_num
-            )
-            self._metrics.ota_image_total_directories_num = (
-                self._ota_metadata.total_dirs_num
-            )
-            self._metrics.ota_image_total_symlinks_num = (
-                self._ota_metadata.total_symlinks_num
-            )
-        except ota_errors.OTAError:
-            raise  # raise top-level OTAError as it
-        except ota_metadata_error.MetadataJWTVerificationFailed as e:
-            _err_msg = f"failed to verify metadata.jwt: {e!r}"
-            logger.error(_err_msg)
-            raise ota_errors.MetadataJWTVerficationFailed(
-                _err_msg, module=__name__
-            ) from e
-        except (ota_metadata_error.MetadataJWTPayloadInvalid, AssertionError) as e:
-            _err_msg = f"metadata.jwt is invalid: {e!r}"
-            logger.error(_err_msg)
-            raise ota_errors.MetadataJWTInvalid(_err_msg, module=__name__) from e
-        except Exception as e:
-            _err_msg = f"failed to prepare ota metafiles: {e!r}"
-            logger.error(_err_msg)
-            raise ota_errors.OTAMetaDownloadFailed(_err_msg, module=__name__) from e
-
-
-class _OTAUpdater(_OTAUpdateOperator):
+class OTAUpdater(OTAUpdateOperator):
     """The implementation of OTA update logic."""
 
     def __init__(
@@ -421,7 +130,7 @@ class _OTAUpdater(_OTAUpdateOperator):
                 start=1,
             ):
                 _now = time.time()
-                if _download_exception_handler(_fut):
+                if download_exception_handler(_fut):
                     err_count, file_size, downloaded_bytes = _fut.result()
 
                     _merged_payload.processed_file_num += 1
@@ -879,552 +588,3 @@ class _OTAUpdater(_OTAUpdateOperator):
         finally:
             ensure_umount(self._session_workdir, ignore_error=True)
             shutil.rmtree(self._session_workdir, ignore_errors=True)
-
-
-class _OTAClientUpdater(_OTAUpdateOperator):
-    """The implementation of OTA client update logic."""
-
-    def __init__(
-        self,
-        client_update_control_flags: ClientUpdateControlFlags,
-        **kwargs,
-    ) -> None:
-        # ------ init base class ------ #
-        super().__init__(**kwargs)
-
-        # --- Event flag to control client update ---- #
-        self.client_update_control_flags = client_update_control_flags
-
-        # ------ setup OTA client package parser ------ #
-        self._ota_client_package = OTAClientPackageDownloader(
-            base_url=self.url_base,
-            ota_metadata=self._ota_metadata,
-            session_dir=self._session_workdir,
-            package_install_dir=cfg.OTACLIENT_INSTALLATION_RELEASE,
-            squashfs_file=cfg.DYNAMIC_CLIENT_SQUASHFS_FILE,
-        )
-
-    def _execute_client_update(self):
-        """Implementation of OTA updating."""
-        logger.info(f"execute local update({ecu_info.ecu_id=}): {self.update_version=}")
-
-        try:
-            self._handle_upper_proxy()
-            self._process_metadata(only_metadata_verification=True)
-            self._download_client_package_resources()
-            self._wait_sub_ecus()
-            if self._is_same_client_package_version():
-                # to notify the status report after reboot
-                _err_msg = "client package version is the same, skip client update"
-                logger.info(_err_msg)
-                raise ota_errors.ClientUpdateSameVersions(
-                    _err_msg,
-                    module=__name__,
-                )
-            else:
-                self._copy_client_package()
-                self._notify_data_ready()
-        except ota_errors.ClientUpdateSameVersions:
-            raise
-        except Exception as e:
-            _err_msg = f"client update failed: {e!r}"
-            logger.warning(_err_msg)
-            raise ota_errors.ClientUpdateFailed(
-                _err_msg,
-                module=__name__,
-            ) from e
-
-    def _download_client_package_resources(self) -> None:
-        """Download OTA client."""
-        logger.info("start to download client manifest and package...")
-        self._status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAUpdatePhaseChangeReport(
-                    new_update_phase=UpdatePhase.DOWNLOADING_OTA_CLIENT,
-                    trigger_timestamp=int(time.time()),
-                ),
-                session_id=self.session_id,
-            )
-        )
-
-        _condition = threading.Condition()
-        try:
-            for _fut in self._download_helper.download_meta_files(
-                self._ota_client_package.download_client_package(_condition),
-                condition=_condition,
-            ):
-                _download_exception_handler(_fut)
-        except Exception as e:
-            logger.error(f"failed to download otaclient package: {e!r}")
-            raise ota_errors.OTAClientPackageDownloadFailed(module=__name__) from e
-        finally:
-            self._downloader_pool.shutdown()
-
-    def _wait_sub_ecus(self) -> None:
-        logger.info("wait for all sub-ECU to finish...")
-        # wait for all sub-ECU to finish OTA update
-        result = wait_and_log(
-            check_flag=self.ecu_status_flags.any_child_ecu_in_update.is_set,
-            check_for=False,
-            message="client updating in sub ecus",
-            log_func=logger.info,
-            timeout=cfg.CLIENT_UPDATE_TIMEOUT,
-        )
-        if result is False:
-            logger.warning("sub-ECU client was aborted, skip waiting for sub-ecus")
-
-    def _is_same_client_package_version(self) -> bool:
-        return self._ota_client_package.is_same_client_package_version()
-
-    def _copy_client_package(self) -> None:
-        self._ota_client_package.copy_client_package()
-
-    def _notify_data_ready(self):
-        """Notify the main process that the client package is ready."""
-        logger.info("notify main process that the client package is ready..")
-        self.client_update_control_flags.notify_data_ready_event.set()
-
-    # API
-
-    def execute(self) -> None:
-        """Main entry for executing local OTA client update."""
-        try:
-            self._execute_client_update()
-        except Exception as e:
-            logger.error(f"client update failed: {e!r}")
-            raise
-        finally:
-            ensure_umount(self._session_workdir, ignore_error=True)
-            shutil.rmtree(self._session_workdir, ignore_errors=True)
-
-
-class _OTARollbacker:
-    def __init__(self, boot_controller: BootControllerProtocol) -> None:
-        self._boot_controller = boot_controller
-
-    def execute(self):
-        try:
-            self._boot_controller.pre_rollback()
-            self._boot_controller.post_rollback()
-            self._boot_controller.finalizing_rollback()
-        except ota_errors.OTAError as e:
-            logger.error(f"rollback failed: {e!r}")
-            self._boot_controller.on_operation_failure()
-            raise
-
-
-class OTAClient:
-    def __init__(
-        self,
-        *,
-        ecu_status_flags: MultipleECUStatusFlags,
-        proxy: Optional[str] = None,
-        status_report_queue: Queue[StatusReport],
-        client_update_control_flags: ClientUpdateControlFlags,
-        shm_metrics_reader: SharedOTAClientMetricsReader,
-    ) -> None:
-        self.my_ecu_id = ecu_info.ecu_id
-        self.proxy = proxy
-        self.ecu_status_flags = ecu_status_flags
-
-        self._status_report_queue = status_report_queue
-        self._client_update_control_flags = client_update_control_flags
-
-        self._shm_metrics_reader = shm_metrics_reader
-        atexit.register(shm_metrics_reader.atexit)
-
-        self._live_ota_status = OTAStatus.INITIALIZED
-        self.started = False
-
-        self._runtime_dir = _runtime_dir = Path(cfg.RUN_DIR)
-        _runtime_dir.mkdir(exist_ok=True, parents=True)
-        self._update_session_dir = _update_session_dir = Path(cfg.RUNTIME_OTA_SESSION)
-
-        # NOTE: for each otaclient instance lifecycle, only one tmpfs will be mounted.
-        #       If otaclient terminates by signal, umounting will be handled by _on_shutdown.
-        #       If otaclient exits on successful OTA, no need to umount it manually as we will reboot soon.
-        ensure_umount(_update_session_dir, ignore_error=True)
-        _update_session_dir.mkdir(exist_ok=True, parents=True)
-        try:
-            ensure_mount(
-                "tmpfs",
-                _update_session_dir,
-                mount_func=partial(
-                    mount_tmpfs, size_in_mb=cfg.SESSION_WD_TMPFS_SIZE_IN_MB
-                ),
-                raise_exception=True,
-            )
-        except Exception as e:
-            logger.warning(f"failed to mount tmpfs for OTA runtime use: {e!r}")
-            logger.warning("will directly use /run tmpfs for OTA runtime!")
-
-        self._metrics = OTAMetricsData()
-        self._metrics.ecu_id = self.my_ecu_id
-        self._metrics.enable_local_ota_proxy_cache = (
-            proxy_info.enable_local_ota_proxy_cache
-        )
-
-        try:
-            _boot_controller_type = get_boot_controller(ecu_info.bootloader)
-        except Exception as e:
-            self._on_failure(
-                e,
-                ota_status=OTAStatus.FAILURE,
-                failure_type=FailureType.UNRECOVERABLE,
-                failure_reason=f"failed to determine boot controller or create_standby mode: {e!r}",
-            )
-            return
-
-        try:
-            self.boot_controller = _boot_controller_type()
-        except Exception as e:
-            self._on_failure(
-                e,
-                ota_status=OTAStatus.FAILURE,
-                failure_type=FailureType.UNRECOVERABLE,
-                failure_reason=f"boot controller startup failed: {e!r}",
-            )
-            return
-        self._metrics.bootloader_type = self.boot_controller.bootloader_type
-
-        # load and report booted OTA status
-        _boot_ctrl_loaded_ota_status = self.boot_controller.get_booted_ota_status()
-        self._live_ota_status = _boot_ctrl_loaded_ota_status
-        self.current_version = self.boot_controller.load_version()
-
-        status_report_queue.put_nowait(
-            StatusReport(
-                payload=SetOTAClientMetaReport(
-                    firmware_version=self.current_version,
-                ),
-            )
-        )
-        status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAStatusChangeReport(
-                    new_ota_status=_boot_ctrl_loaded_ota_status,
-                ),
-            )
-        )
-        self._metrics.current_firmware_version = self.current_version
-
-        self.ca_chains_store = None
-        try:
-            self.ca_chains_store = load_ca_cert_chains(cfg.CERT_DPATH)
-        except CACertStoreInvalid as e:
-            _err_msg = f"failed to import ca_chains_store: {e!r}, OTA will NOT occur on no CA chains installed!!!"
-            logger.error(_err_msg)
-
-            self.ca_chains_store = CAChainStore()
-
-        self.started = True
-        logger.info("otaclient started")
-
-        # NOTE: not doing fstrim at startup when running as dynamic otaclient
-        if not _env.is_dynamic_client_running() and cfg.FSTRIM_AT_OTACLIENT_STARTUP:
-            logger.info(
-                "spawn a subprocess to do fstrim on active slot"
-                f"(timeout={cfg.FSTRIM_AT_OTACLIENT_STARTUP_TIMEOUT}s)"
-            )
-            fstrim_at_subprocess(
-                Path(CANONICAL_ROOT),
-                wait=False,
-                timeout=cfg.FSTRIM_AT_OTACLIENT_STARTUP_TIMEOUT,
-            )
-
-    def _on_failure(
-        self,
-        exc: Exception,
-        *,
-        ota_status: OTAStatus,
-        failure_reason: str,
-        failure_type: FailureType,
-    ) -> None:
-        try:
-            _traceback = get_traceback(exc)
-
-            logger.error(failure_reason)
-            logger.error(f"last error traceback: \n{_traceback}")
-
-            self._status_report_queue.put_nowait(
-                StatusReport(
-                    payload=OTAStatusChangeReport(
-                        new_ota_status=ota_status,
-                        failure_type=failure_type,
-                        failure_reason=failure_reason,
-                        failure_traceback=_traceback,
-                    ),
-                )
-            )
-            self._metrics.failure_type = failure_type
-            self._metrics.failure_reason = failure_reason
-            self._metrics.failed_status = ota_status
-        finally:
-            del exc  # prevent ref cycle
-
-    def _exit_from_dynamic_client(self) -> None:
-        """Exit from dynamic client."""
-        if not _env.is_dynamic_client_running():
-            # dynamic client is not running, no need to exit
-            return
-
-        logger.info("exit from dynamic client...")
-        self._client_update_control_flags.request_shutdown_event.set()
-
-    # API
-
-    @property
-    def live_ota_status(self) -> OTAStatus:
-        return self._live_ota_status
-
-    @property
-    def is_busy(self) -> bool:
-        return self._live_ota_status in [
-            OTAStatus.UPDATING,
-            OTAStatus.ROLLBACKING,
-            OTAStatus.CLIENT_UPDATING,
-        ]
-
-    def update(self, request: UpdateRequestV2) -> None:
-        """
-        NOTE that update API will not raise any exceptions. The failure information
-            is available via status API.
-        """
-        self._live_ota_status = OTAStatus.UPDATING
-        request_id = request.request_id
-        new_session_id = request.session_id
-        self._status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAStatusChangeReport(
-                    new_ota_status=OTAStatus.UPDATING,
-                ),
-                session_id=new_session_id,
-            )
-        )
-        logger.info(
-            f"start new OTA update request:{request_id}, session: {new_session_id=}"
-        )
-
-        session_wd = self._update_session_dir / new_session_id
-        self._metrics.request_id = request_id
-        self._metrics.session_id = new_session_id
-        try:
-            logger.info("[update] entering local update...")
-            if not self.ca_chains_store:
-                raise ota_errors.MetadataJWTVerficationFailed(
-                    "no CA chains are installed, reject any OTA update",
-                    module=__name__,
-                )
-
-            _OTAUpdater(
-                version=request.version,
-                raw_url_base=request.url_base,
-                cookies_json=request.cookies_json,
-                session_wd=session_wd,
-                ca_chains_store=self.ca_chains_store,
-                boot_controller=self.boot_controller,
-                ecu_status_flags=self.ecu_status_flags,
-                upper_otaproxy=self.proxy,
-                status_report_queue=self._status_report_queue,
-                session_id=new_session_id,
-                metrics=self._metrics,
-                shm_metrics_reader=self._shm_metrics_reader,
-            ).execute()
-        except ota_errors.OTAError as e:
-            self._live_ota_status = OTAStatus.FAILURE
-            self._on_failure(
-                e,
-                ota_status=OTAStatus.FAILURE,
-                failure_reason=e.get_failure_reason(),
-                failure_type=e.failure_type,
-            )
-            self._exit_from_dynamic_client()
-        finally:
-            shutil.rmtree(session_wd, ignore_errors=True)
-            try:
-                if self._shm_metrics_reader:
-                    _shm_metrics = self._shm_metrics_reader.sync_msg()
-                    self._metrics.shm_merge(_shm_metrics)
-            except Exception as e:
-                logger.error(f"failed to merge metrics: {e!r}")
-            self._metrics.publish()
-
-    def client_update(self, request: ClientUpdateRequestV2) -> None:
-        """
-        NOTE that client update API will not raise any exceptions. The failure information
-            is available via status API.
-        """
-        if _env.is_dynamic_client_running():
-            # Duplicates client update should not be allowed.
-            # TODO(airkei) [2025-06-19]: should return the dedicated error code for "client update"
-            return
-
-        self._live_ota_status = OTAStatus.CLIENT_UPDATING
-        request_id = request.request_id
-        new_session_id = request.session_id
-        self._status_report_queue.put_nowait(
-            StatusReport(
-                payload=OTAStatusChangeReport(
-                    new_ota_status=OTAStatus.CLIENT_UPDATING,
-                ),
-                session_id=new_session_id,
-            )
-        )
-        logger.info(
-            f"start new OTA client update request: {request_id}, session: {new_session_id=}"
-        )
-
-        session_wd = self._update_session_dir / new_session_id
-        try:
-            logger.info("[client update] entering local update...")
-            if not self.ca_chains_store:
-                raise ota_errors.MetadataJWTVerficationFailed(
-                    "no CA chains are installed, reject any OTA update",
-                    module=__name__,
-                )
-
-            _OTAClientUpdater(
-                version=request.version,
-                raw_url_base=request.url_base,
-                cookies_json=request.cookies_json,
-                session_wd=session_wd,
-                ca_chains_store=self.ca_chains_store,
-                ecu_status_flags=self.ecu_status_flags,
-                upper_otaproxy=self.proxy,
-                status_report_queue=self._status_report_queue,
-                session_id=new_session_id,
-                client_update_control_flags=self._client_update_control_flags,
-                metrics=self._metrics,
-                shm_metrics_reader=self._shm_metrics_reader,
-            ).execute()
-        except ota_errors.OTAError:
-            logger.warning("client update failed")
-            # TODO(airkei) [2025-06-19]: should return the dedicated error code for "client update"
-            # As temporary workaround, we set the status to SUCCESS here when current process is dynamic client.
-            self._live_ota_status = OTAStatus.SUCCESS
-            self._status_report_queue.put_nowait(
-                StatusReport(
-                    payload=OTAStatusChangeReport(
-                        new_ota_status=OTAStatus.SUCCESS,
-                    ),
-                    session_id=new_session_id,
-                )
-            )
-        except Exception:
-            self._client_update_control_flags.request_shutdown_event.set()
-        finally:
-            shutil.rmtree(session_wd, ignore_errors=True)
-
-    def main(
-        self,
-        *,
-        req_queue: mp_queue.Queue[IPCRequest],
-        resp_queue: mp_queue.Queue[IPCResponse],
-    ) -> NoReturn:
-        """Main loop of ota_core process."""
-        _allow_request_after = 0
-        while True:
-            _now = int(time.time())
-            try:
-                request = req_queue.get(timeout=OP_CHECK_INTERVAL)
-            except Empty:
-                continue
-
-            if _now < _allow_request_after or self.is_busy:
-                _err_msg = (
-                    f"otaclient is busy at {self._live_ota_status} or "
-                    f"request too quickly({_allow_request_after=}), "
-                    f"reject {request}"
-                )
-                logger.warning(_err_msg)
-                resp_queue.put_nowait(
-                    IPCResponse(
-                        res=IPCResEnum.REJECT_BUSY,
-                        msg=_err_msg,
-                        session_id=request.session_id,
-                    )
-                )
-
-            elif isinstance(request, UpdateRequestV2):
-                _update_thread = threading.Thread(
-                    target=self.update,
-                    args=[request],
-                    daemon=True,
-                    name="ota_update_executor",
-                )
-                _update_thread.start()
-
-                resp_queue.put_nowait(
-                    IPCResponse(
-                        res=IPCResEnum.ACCEPT,
-                        session_id=request.session_id,
-                    )
-                )
-                _allow_request_after = _now + HOLD_REQ_HANDLING_ON_ACK_REQUEST
-
-            elif isinstance(request, ClientUpdateRequestV2):
-                _client_update_thread = threading.Thread(
-                    target=self.client_update,
-                    args=[request],
-                    daemon=True,
-                    name="ota_client_update_executor",
-                )
-                _client_update_thread.start()
-
-                resp_queue.put_nowait(
-                    IPCResponse(
-                        res=IPCResEnum.ACCEPT,
-                        session_id=request.session_id,
-                    )
-                )
-                _allow_request_after = (
-                    _now + HOLD_REQ_HANDLING_ON_ACK_CLIENT_UPDATE_REQUEST
-                )
-            else:
-                _err_msg = f"request is invalid: {request=}, {self._live_ota_status=}"
-                logger.error(_err_msg)
-                resp_queue.put_nowait(
-                    IPCResponse(
-                        res=IPCResEnum.REJECT_OTHER,
-                        msg=_err_msg,
-                        session_id=request.session_id,
-                    )
-                )
-
-
-def ota_core_process(
-    *,
-    shm_writer_factory: Callable[[], SharedOTAClientStatusWriter],
-    shm_metrics_reader_factory: Callable[[], SharedOTAClientMetricsReader],
-    ecu_status_flags: MultipleECUStatusFlags,
-    op_queue: mp_queue.Queue[IPCRequest],
-    resp_queue: mp_queue.Queue[IPCResponse],
-    max_traceback_size: int,  # in bytes
-    client_update_control_flags: ClientUpdateControlFlags,
-):
-    from otaclient._logging import configure_logging
-    from otaclient.configs.cfg import proxy_info
-    from otaclient.ota_core import OTAClient
-
-    configure_logging()
-
-    shm_writer = shm_writer_factory()
-    shm_metrics_reader = shm_metrics_reader_factory()
-
-    _local_status_report_queue = Queue()
-    _status_monitor = OTAClientStatusCollector(
-        msg_queue=_local_status_report_queue,
-        shm_status=shm_writer,
-        max_traceback_size=max_traceback_size,
-    )
-    _status_monitor.start()
-    _status_monitor.start_log_thread()
-
-    _ota_core = OTAClient(
-        ecu_status_flags=ecu_status_flags,
-        proxy=proxy_info.get_proxy_for_local_ota(),
-        status_report_queue=_local_status_report_queue,
-        client_update_control_flags=client_update_control_flags,
-        shm_metrics_reader=shm_metrics_reader,
-    )
-    _ota_core.main(req_queue=op_queue, resp_queue=resp_queue)
